requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[argon2]
pydantic
pydantic[email]
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
alembic
langchain
langchain-core
langchain-community
langchain-qdrant
langchain-google-genai
langchain-huggingface
langgraph
qdrant-client
sentence-transformers
flashrank
unstructured[pdf]
pypdf
pdf2image
lxml
beautifulsoup4
mlflow
deepeval
pytest
requests
py-cpuinfo
pymupdf4llm
pdfplumber
google-generativeai
google-genai